# tich-hop-tiktok-bitrix24/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/webhooks/tiktok/leads": {
      "post": {
        "operationId": "TikTokController_handleLeadWebhook",
        "summary": "Receive TikTok Lead Generation webhook",
        "description": "Receives lead events from TikTok Lead Generation Forms, stores raw data, and queues for async processing.",
        "parameters": [
          {
            "name": "TikTok-Signature",
            "in": "header",
            "description": "HMAC-SHA256 signature of request payload",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook received and queued successfully"
          },
          "401": {
            "description": "Invalid or missing signature"
          }
        },
        "tags": [
          "Webhooks"
        ]
      }
    },
...
    "/api/v1/leads": {
      "get": {
        "operationId": "LeadsController_getLeads",
        "summary": "Get paginated list of leads",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "source",
            "required": false,
            "in": "query",
            "schema": {
              "example": "tiktok",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return paginated leads"
          }
        },
        "tags": [
          "Management"
        ]
      }
    },
    "/api/v1/leads/{id}/convert-to-deal": {
      "post": {
        "operationId": "LeadsController_convertToDeal",
        "summary": "Manually convert a lead to deal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lead successfully converted to deal"
          },
          "404": {
            "description": "Lead not found"
          }
        },
        "tags": [
          "Management"
        ]
      }
    },
...
    "/api/v1/reports/export": {
      "get": {
        "operationId": "AnalyticsController_exportReport",
        "summary": "Export lead and conversion report in CSV (with UTF-8 BOM) or JSON format",
        "parameters": [
          {
            "name": "format",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "csv",
                "json"
              ],
              "type": "string"
            }
          },
          {
            "name": "date_range",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "7d",
                "30d",
                "90d",
                "all"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Download CSV file or get JSON report"
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
...
  },
  "info": {
    "title": "TikTok to Bitrix24 Integration API",
    "description": "API documentation for TikTok Lead Generation & Bitrix24 CRM Synchronization",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
...
}